LayoutTests/http/tests/security/credentials-main-resource.html

<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
    internals.settings.setStorageBlockingPolicy('BlockThirdParty');
}
window.location = "http://testuser:testpass@127.0.0.1:8000/security/resources/credentials-main-resource.php";
</script>
