LayoutTests/http/tests/security/credentials-main-resource-expected.txt

ALERT: Authenticated as user: testuser password: testpass
Main Resource Credentials: testuser, testpass
